requirements.txt
datasets
matplotlib
numpy
nvgpu
pillow
requests
torch
torchserve
torch_tb_profiler
tqdm
transformers